Fix missing space between -i and ''
The --i option in sed is used for in-place ediiting of files. It should
be followed by an optional argument which is the extension of the backup
file to be created before editing. If there's no space between -i and
'' then a new file gets created, which isn't the intention of the
command.