Fix errors in context.py
We ask ruff to ignore E402 (module level import not at top of file) and
F403 (unable to detect undefined names) otherwise we get errors when we
run `make lint` or `make format`. Also the github workflow fails when
the code is checked in. There's just no need.

We also make PROJECT_ROOT upper case which is a pattern for variables
that never change (like constants but Python doesn't have a dedicated
syntax for defining constants). The idea is to prevent programmers from
reassigning a value that holds a constant andf the uppercase letters is
a known and agree naming convention. So why don't we just use it eh ;)